*It is result of many weeks of episodic trying improve code and add support of Return, function arguments. Modifications: * Calculations implemented on stack * Added Return Token and few tokens around it * Added few utils for DFS * Added algorithm for creating CFG from AST * Fixed some bugs * Interface ISyntaxNode expanded with few important methods
This text writed only by memory. Some changes not described, but mainline is given.

+++ src/syntax_nodes/terminals/print_syntax_node.cpp
@@ -19,7 +19,7 @@ bool PrintSyntaxNode::compare( const ISyntaxNode& node ) const
    {
       if( node->Children().size() != this->Children().size() )
          return;
-      for( int i = 0; i < Children().size(); ++i )
+      for( size_t i = 0; i < Children().size(); ++i )
       {
          const auto& lft_child = ( *this )[ i ];
          const auto& rht_child = ( *node )[ i ];
